There was a mix-up in the output ranges for some noise.
See FastNoiseLite for the small changes. The rest of the changes were things I needed to get this to build on Windows 10, like creating intermediate directories before writing a file into a nested directory. CommonUtil.toList()... I have no idea how it compiled before. As far as I can tell, it returns a List of Object, which would need to be cast whenever it is used... And it isn't being cast... I switched that to Arrays.toList(), which would be used in either case. I'm not sure how the versioning in the Maven pom.xml files works, but I needed to drop a child's version by 1 to match the parent.

The run configurations are here so I don't have to create them again, but it's no problem if you want to erase them. I'm not used to seeing the .idea folder in repos very often.